requirements.txt
chromadb==0.4.18
ctransformers==0.2.27
InstructorEmbedding==1.0.1
langchain==0.1.0
librosa==0.10.1
llama_cpp_python==0.2.20
pypdfium2==4.24.0
sentence-transformers==2.2.2
PyYAML==6.0.1
torch==2.1.1
streamlit==1.28.2
streamlit-mic-recorder==0.0.4
transformers==4.35.2